[5994] - don't run create degree if no degrees (#3739)

+++ app/services/degrees/create_from_csv_row.rb
@@ -14,6 +14,8 @@ module Degrees
     end
 
     def call
+      return if country.blank?
+
       trainee.transaction do
         trainee.degrees.destroy_all
         trainee.degrees.create!(mapped_degree_attributes)
